requirements.txt
google-genai
pillow
pydantic
python-dotenv
python-dotenv
